# frontend/backend-openapi.json
{
    "openapi": "3.1.0",
    "info": { "title": "FastAPI", "version": "0.1.0" },
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": ["auth"],
                "summary": "Register User",
                "operationId": "register_user_api_v1_auth_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": ["auth"],
                "summary": "Login User",
                "operationId": "login_user_api_v1_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLogin"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": ["auth"],
                "summary": "Read Users Me",
                "operationId": "read_users_me_api_v1_auth_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [{ "HTTPBearer": [] }]
            }
        },
        "/api/v1/chat": {
            "get": {
                "tags": ["chat"],
                "summary": "Get All Chats",
                "operationId": "get_all_chats_api_v1_chat_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ChatResponse"
                                    },
                                    "type": "array",
                                    "title": "Response Get All Chats Api V1 Chat Get"
                                }
                            }
                        }
                    }
                },
                "security": [{ "HTTPBearer": [] }]
            },
            "post": {
                "tags": ["chat"],
                "summary": "Create Chat",
                "operationId": "create_chat_api_v1_chat_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromptID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpandedChatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [{ "HTTPBearer": [] }]
            }
        },
        "/api/v1/chat/{chat_id}": {
            "get": {
                "tags": ["chat"],
                "summary": "Get Chat",
                "operationId": "get_chat_api_v1_chat__chat_id__get",
                "security": [{ "HTTPBearer": [] }],
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Chat Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpandedChatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["chat"],
                "summary": "Delete Chat",
                "operationId": "delete_chat_api_v1_chat__chat_id__delete",
                "security": [{ "HTTPBearer": [] }],
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Chat Id" }
                    }
                ],
                "responses": {
                    "204": { "description": "Successful Response" },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/chat/{chat_id}/message": {
            "post": {
                "tags": ["chat"],
                "summary": "Create Message",
                "operationId": "create_message_api_v1_chat__chat_id__message_post",
                "security": [{ "HTTPBearer": [] }],
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Chat Id" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BaseMessage"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/document": {
            "get": {
                "tags": ["document"],
                "summary": "Get Documents For User",
                "operationId": "get_documents_for_user_api_v1_document_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/DocumentResponse"
                                    },
                                    "type": "array",
                                    "title": "Response Get Documents For User Api V1 Document Get"
                                }
                            }
                        }
                    }
                },
                "security": [{ "HTTPBearer": [] }]
            },
            "post": {
                "tags": ["document"],
                "summary": "Upload Document",
                "operationId": "upload_document_api_v1_document_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_document_api_v1_document_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [{ "HTTPBearer": [] }]
            }
        },
        "/api/v1/document/{document_id}": {
            "get": {
                "tags": ["document"],
                "summary": "Get Document",
                "operationId": "get_document_api_v1_document__document_id__get",
                "security": [{ "HTTPBearer": [] }],
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Document Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpandedDocumentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["document"],
                "summary": "Delete Document",
                "operationId": "delete_document_api_v1_document__document_id__delete",
                "security": [{ "HTTPBearer": [] }],
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Document Id" }
                    }
                ],
                "responses": {
                    "204": { "description": "Successful Response" },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BaseMessage": {
                "properties": {
                    "content": { "type": "string", "title": "Content" },
                    "documents_ids": {
                        "items": { "type": "integer" },
                        "type": "array",
                        "title": "Documents Ids",
                        "default": []
                    }
                },
                "type": "object",
                "required": ["content"],
                "title": "BaseMessage"
            },
            "Body_upload_document_api_v1_document_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": ["file"],
                "title": "Body_upload_document_api_v1_document_post"
            },
            "ChatResponse": {
                "properties": {
                    "chat_id": { "type": "integer", "title": "Chat Id" },
                    "is_active": { "type": "boolean", "title": "Is Active" },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": ["chat_id", "is_active", "created_at"],
                "title": "ChatResponse"
            },
            "DocumentResponse": {
                "properties": {
                    "document_id": {
                        "type": "integer",
                        "title": "Document Id"
                    },
                    "minio_url": { "type": "string", "title": "Minio Url" },
                    "filename": { "type": "string", "title": "Filename" },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "is_general": { "type": "boolean", "title": "Is General" }
                },
                "type": "object",
                "required": [
                    "document_id",
                    "minio_url",
                    "filename",
                    "created_at",
                    "is_general"
                ],
                "title": "DocumentResponse"
            },
            "ExpandedChatResponse": {
                "properties": {
                    "chat_id": { "type": "integer", "title": "Chat Id" },
                    "is_active": { "type": "boolean", "title": "Is Active" },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "prompt": { "type": "string", "title": "Prompt" },
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/MessageResponse"
                        },
                        "type": "array",
                        "title": "Messages",
                        "default": []
                    }
                },
                "type": "object",
                "required": ["chat_id", "is_active", "created_at", "prompt"],
                "title": "ExpandedChatResponse"
            },
            "ExpandedDocumentResponse": {
                "properties": {
                    "document_id": {
                        "type": "integer",
                        "title": "Document Id"
                    },
                    "minio_url": { "type": "string", "title": "Minio Url" },
                    "filename": { "type": "string", "title": "Filename" },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "is_general": { "type": "boolean", "title": "Is General" },
                    "document_length": {
                        "type": "integer",
                        "title": "Document Length"
                    }
                },
                "type": "object",
                "required": [
                    "document_id",
                    "minio_url",
                    "filename",
                    "created_at",
                    "is_general",
                    "document_length"
                ],
                "title": "ExpandedDocumentResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "MessageResponse": {
                "properties": {
                    "content": { "type": "string", "title": "Content" },
                    "documents_ids": {
                        "items": { "type": "integer" },
                        "type": "array",
                        "title": "Documents Ids"
                    },
                    "message_id": { "type": "integer", "title": "Message Id" },
                    "message_type": {
                        "$ref": "#/components/schemas/MessageType"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "hidden_comments": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Hidden Comments"
                    }
                },
                "type": "object",
                "required": [
                    "content",
                    "documents_ids",
                    "message_id",
                    "message_type",
                    "created_at",
                    "hidden_comments"
                ],
                "title": "MessageResponse"
            },
            "MessageType": {
                "type": "integer",
                "enum": [1, 0],
                "title": "MessageType"
            },
            "PromptID": {
                "properties": {
                    "prompt_id": { "type": "integer", "title": "Prompt Id" }
                },
                "type": "object",
                "required": ["prompt_id"],
                "title": "PromptID"
            },
            "Role": { "type": "integer", "enum": [1, 0], "title": "Role" },
            "Token": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "default": "bearer"
                    }
                },
                "type": "object",
                "required": ["access_token"],
                "title": "Token"
            },
            "UserCreate": {
                "properties": {
                    "username": { "type": "string", "title": "Username" },
                    "email": { "type": "string", "title": "Email" },
                    "role": {
                        "$ref": "#/components/schemas/Role",
                        "default": 1
                    },
                    "password": { "type": "string", "title": "Password" }
                },
                "type": "object",
                "required": ["username", "email", "password"],
                "title": "UserCreate"
            },
            "UserLogin": {
                "properties": {
                    "email": { "type": "string", "title": "Email" },
                    "password": { "type": "string", "title": "Password" }
                },
                "type": "object",
                "required": ["email", "password"],
                "title": "UserLogin"
            },
            "UserResponse": {
                "properties": {
                    "username": { "type": "string", "title": "Username" },
                    "email": { "type": "string", "title": "Email" },
                    "role": {
                        "$ref": "#/components/schemas/Role",
                        "default": 1
                    },
                    "id": { "type": "integer", "title": "Id" }
                },
                "type": "object",
                "required": ["username", "email", "id"],
                "title": "UserResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                { "type": "string" },
                                { "type": "integer" }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": { "type": "string", "title": "Message" },
                    "type": { "type": "string", "title": "Error Type" }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": { "type": "http", "scheme": "bearer" }
        }
    }
}